requirements.txt
# Core GUI
PyQt6>=6.6.0
PyQt6-Qt6>=6.6.0

# Image processing & duplicate detection
Pillow>=10.0.0
imagehash>=4.3.1
opencv-python>=4.8.0

# Document processing
python-docx>=1.0.0
odfpy>=1.4.1
PyPDF2>=3.0.1
pytesseract>=0.3.10  # OCR for PDFs
pdfplumber>=0.10.0

# Video processing
moviepy>=1.0.3

# Text similarity
python-Levenshtein>=0.21.0
rapidfuzz>=3.0.0

# Database
sqlalchemy>=2.0.0

# File handling
python-magic>=0.4.27
watchdog>=3.0.0

# Optional: ML for image categorization
# requests>=2.31.0  # for Ollama API calls
